完成网关、分布式事务、微服务监控

+++ common/common-core/src/main/java/com/zclcs/common/core/base/BasePageAo.java
@@ -20,14 +20,14 @@ public class BasePageAo {
      */
     @NotNull(message = "请输入页码")
     @ApiModelProperty(value = "页码", required = true)
-    private Long pageNum;
+    private Integer pageNum;
 
     /**
      * 页数
      */
     @NotNull(message = "请输入页数")
     @ApiModelProperty(value = "页数", required = true)
-    private Long pageSize;
+    private Integer pageSize;
 
     /**
      * 排序字段
